Add options for autoclaim
On servers with poor hardware, it can take a long time to spawn a mission. This update captures the player's name that started the mission and offers the claim to that player. This prevents another player from sneaking in front of that player and getting the claim unfairly. I also updated the warning system so that the player object gets removed from the warning array if the player leaves the area. That way the player gets the warning message again if he/she reenters the area.